@@ -1,11 +1,11 @@
 name: flutter_call_kit
-description: Flutter iOS CallKit framework - Currently iOS >= 10.0 only
-version: 0.0.3
-author: Onyemaechi Okafor <peerwaya@yahoo.com>
-homepage: https://github.com/peerwaya/flutter_call_kit
+description: A new Flutter plugin.
+version: 0.0.1
+homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
@@ -21,17 +21,21 @@ dev_dependencies:
 # The following section is specific to Flutter.
 flutter:
   # This section identifies this Flutter project as a plugin project.
-  # The androidPackage and pluginClass identifiers should not ordinarily
+  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
-    androidPackage: com.example.flutter_call_kit
-    pluginClass: FlutterCallKitPlugin
+    platforms:
+      android:
+        package: com.example.flutter_call_kit
+        pluginClass: FlutterCallKitPlugin
+      ios:
+        pluginClass: FlutterCallKitPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
   #
   # For details regarding assets in packages, see
   # https://flutter.dev/assets-and-images/#from-packages
