@@ -89,28 +89,22 @@ class IOSOptions {
   final bool includesCallsInRecents;
 
   IOSOptions(
-    this.appName, {
-    this.imageName = "",
-    this.ringtoneSound = "",
-    this.maximumCallGroups = 3,
-    this.maximumCallsPerCallGroup = 1,
-    this.supportsVideo = true,
-    this.includesCallsInRecents = true,
-  })  : assert(appName != null),
-        assert(imageName != null),
-        assert(ringtoneSound != null),
-        assert(maximumCallGroups != null),
-        assert(maximumCallsPerCallGroup != null),
-        assert(supportsVideo != null),
-        assert(includesCallsInRecents != null);
+      this.appName, {
+        this.imageName = "",
+        this.ringtoneSound = "",
+        this.maximumCallGroups = 3,
+        this.maximumCallsPerCallGroup = 1,
+        this.supportsVideo = false,
+        this.includesCallsInRecents = true,
+      });
 
   Map<String, dynamic> toMap() {
     return {
       "appName": appName,
       "imageName": imageName,
       "ringtoneSound": ringtoneSound,
-      "maximumCallGroups": maximumCallGroups?.toString(),
-      "maximumCallsPerCallGroup": maximumCallsPerCallGroup?.toString(),
+      "maximumCallGroups": maximumCallGroups.toString(),
+      "maximumCallsPerCallGroup": maximumCallsPerCallGroup.toString(),
       "supportsVideo": supportsVideo,
       "includesCallsInRecents": includesCallsInRecents,
     };
@@ -128,43 +122,43 @@ class FlutterCallKit {
 
   final MethodChannel _channel;
 
-  OnReceiveStartCallAction _didReceiveStartCallAction;
+  OnReceiveStartCallAction? _didReceiveStartCallAction;
 
   /// this means something big changed, so tell the Dart side. The Dart side should
   /// probably respond by hanging up all calls.
-  VoidCallback _onProviderReset;
+  VoidCallback? _onProviderReset;
 
-  OnAnswerCallAction _performAnswerCallAction;
+  OnAnswerCallAction? _performAnswerCallAction;
 
-  OnEndCallAction _performEndCallAction;
+  OnEndCallAction? _performEndCallAction;
 
-  OnActivateAudioSession _didActivateAudioSession;
+  OnActivateAudioSession? _didActivateAudioSession;
 
-  OnDeactivateAudioSession _didDeactivateAudioSession;
+  OnDeactivateAudioSession? _didDeactivateAudioSession;
 
-  OnIncomingCall _didDisplayIncomingCall;
+  OnIncomingCall? _didDisplayIncomingCall;
 
-  OnMuted _didPerformSetMutedCallAction;
-  OnHold _didToggleHoldAction;
-  OnDTMF _didPerformDTMFAction;
+  OnMuted? _didPerformSetMutedCallAction;
+  OnHold? _didToggleHoldAction;
+  OnDTMF? _didPerformDTMFAction;
 
-  OnStartCall _handleStartCallNotification;
+  OnStartCall? _handleStartCallNotification;
 
   /// Configures with [options] and sets up handlers for incoming messages.
   void configure(
-    IOSOptions options, {
-    OnReceiveStartCallAction didReceiveStartCallAction,
-    VoidCallback onProviderReset,
-    OnAnswerCallAction performAnswerCallAction,
-    OnEndCallAction performEndCallAction,
-    OnActivateAudioSession didActivateAudioSession,
-    OnDeactivateAudioSession didDeactivateAudioSession,
-    OnIncomingCall didDisplayIncomingCall,
-    OnMuted didPerformSetMutedCallAction,
-    OnDTMF didPerformDTMFAction,
-    OnHold didToggleHoldAction,
-    OnStartCall handleStartCallNotification,
-  }) {
+      IOSOptions options, {
+        OnReceiveStartCallAction? didReceiveStartCallAction,
+        VoidCallback? onProviderReset,
+        OnAnswerCallAction? performAnswerCallAction,
+        OnEndCallAction? performEndCallAction,
+        OnActivateAudioSession? didActivateAudioSession,
+        OnDeactivateAudioSession? didDeactivateAudioSession,
+        OnIncomingCall? didDisplayIncomingCall,
+        OnMuted? didPerformSetMutedCallAction,
+        OnDTMF? didPerformDTMFAction,
+        OnHold? didToggleHoldAction,
+        OnStartCall? handleStartCallNotification,
+      }) {
     if (!Platform.isIOS) {
       return;
     }
@@ -190,66 +184,66 @@ class FlutterCallKit {
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _didReceiveStartCallAction(map["callUUID"], map["handle"]);
+        return _didReceiveStartCallAction!(map["callUUID"], map["handle"]);
       case "onProviderReset":
         if (_onProviderReset == null) {
           return null;
         }
-        return _onProviderReset();
+        return _onProviderReset!();
       case "performAnswerCallAction":
         if (_performAnswerCallAction == null) {
           return null;
         }
-        return _performAnswerCallAction(
+        return _performAnswerCallAction!(
             call.arguments.cast<String, dynamic>()["callUUID"]);
       case "performEndCallAction":
         if (_performEndCallAction == null) {
           return null;
         }
-        return _performEndCallAction(
+        return _performEndCallAction!(
             call.arguments.cast<String, dynamic>()["callUUID"]);
       case "didActivateAudioSession":
         if (_didActivateAudioSession == null) {
           return null;
         }
-        return _didActivateAudioSession();
+        return _didActivateAudioSession!();
       case "didDeactivateAudioSession":
         if (_didDeactivateAudioSession == null) {
           return null;
         }
-        return _didDeactivateAudioSession();
+        return _didDeactivateAudioSession!();
       case "didDisplayIncomingCall":
         if (_didDisplayIncomingCall == null) {
           print("_didDisplayIncomingCall is null");
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _didDisplayIncomingCall(map["error"], map["callUUID"],
+        return _didDisplayIncomingCall!(map["error"], map["callUUID"],
             map["handle"], map["localizedCallerName"], map["fromPushKit"]);
       case "didPerformSetMutedCallAction":
         if (_didPerformSetMutedCallAction == null) {
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _didPerformSetMutedCallAction(map["muted"], map["callUUID"]);
+        return _didPerformSetMutedCallAction!(map["muted"], map["callUUID"]);
       case "didPerformDTMFAction":
         if (_didPerformDTMFAction == null) {
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _didPerformDTMFAction(map["digits"], map["callUUID"]);
+        return _didPerformDTMFAction!(map["digits"], map["callUUID"]);
       case "didToggleHoldAction":
         if (_didToggleHoldAction == null) {
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _didToggleHoldAction(map["hold"], map["callUUID"]);
+        return _didToggleHoldAction!(map["hold"], map["callUUID"]);
       case "handleStartCallNotification":
         if (_handleStartCallNotification == null) {
           return null;
         }
         Map map = call.arguments.cast<String, dynamic>();
-        return _handleStartCallNotification(map["handle"], map["video"]);
+        return _handleStartCallNotification!(map["handle"], map["video"]);
       default:
         throw UnsupportedError("Unrecognized JSON message");
     }
@@ -264,7 +258,7 @@ class FlutterCallKit {
   Future<void> displayIncomingCall(
       String uuid, String handle, String localizedCallerName,
       {HandleType handleType = HandleType.phoneNumber,
-      bool video = false}) async {
+        bool video = false}) async {
     if (!Platform.isIOS) {
       return;
     }
@@ -287,7 +281,7 @@ class FlutterCallKit {
   ///
   Future<void> startCall(String uuid, String handle, String contactIdentifier,
       {HandleType handleType = HandleType.phoneNumber,
-      bool video = false}) async {
+        bool video = false}) async {
     if (!Platform.isIOS) {
       return;
     }
@@ -376,7 +370,7 @@ class FlutterCallKit {
   ///
   Future<bool> checkIfBusy() async {
     if (!Platform.isIOS) {
-      return null;
+      return false;
     }
     return await _channel.invokeMethod<void>('checkIfBusy') as bool;
   }
@@ -385,7 +379,7 @@ class FlutterCallKit {
   ///
   Future<bool> checkSpeaker() async {
     if (!Platform.isIOS) {
-      return null;
+      return false;
     }
     return await _channel.invokeMethod<void>('checkSpeaker') as bool;
   }
